requirements.txt
Flask==2.3.2
flask-restx==1.1.0
python-dotenv==1.0.0